feat: update Topbar
