+++ apps/console/src/components/index.ts
@@ -1,9 +1,8 @@
 export * from "./AuthPageBase";
 export * from "./ChangePasswordForm";
 export * from "./ConsoleCorePageHead";
-export * from "./Sidebar";
-export * from "./ProfileSidebar";
-export * from "./Topbar";
+export * from "./profile-side-bar";
+export * from "./top-bar";
 export * from "./ErrorBoundary";
 export * from "./LoginForm";
 export * from "./ModelReadmeMarkdown";
