+++ apps/console/src/components/top-bar/index.ts
