+++ packages/design-system/src/new-ui/ComplicateIcons/GitHub.tsx
@@ -1,0 +1,30 @@
+import * as React from "react";
+import {
+  ComplicateIconBase,
+  ComplicateIconBaseProps,
+} from "./ComplicateIconBase";
+
+export const GitHub = React.forwardRef<
+  SVGSVGElement,
+  Omit<ComplicateIconBaseProps, "viewBox" | "children"> & {
+    fillAreaColor: string;
+  }
+>((props, ref) => {
+  const { className, fillAreaColor, ...passThrough } = props;
+  return (
+    <ComplicateIconBase
+      {...passThrough}
+      ref={ref}
+      viewBox="0 0 64 64"
+      className={className}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M32.0003 1C14.3293 1 0 15.3267 0 33.0003C0 47.1388 9.16905 59.1339 21.8837 63.3652C23.483 63.6614 24.0702 62.671 24.0702 61.8258C24.0702 61.0628 24.0405 58.542 24.0267 55.868C15.1242 57.8038 13.2457 52.0924 13.2457 52.0924C11.79 48.3937 9.69259 47.4101 9.69259 47.4101C6.78921 45.424 9.91145 45.4648 9.91145 45.4648C13.1248 45.6906 14.8168 48.7625 14.8168 48.7625C17.6709 53.6546 22.3029 52.2403 24.129 51.4226C24.4162 49.3544 25.2455 47.9427 26.1606 47.1436C19.0529 46.3344 11.5812 43.5905 11.5812 31.3289C11.5812 27.8352 12.8313 24.9806 14.8783 22.7396C14.5461 21.9336 13.4507 18.6789 15.1883 14.2711C15.1883 14.2711 17.8755 13.411 23.9907 17.5512C26.5432 16.8422 29.2808 16.4866 32.0003 16.4745C34.7198 16.4866 37.4594 16.8422 40.0168 17.5512C46.1245 13.411 48.808 14.2711 48.808 14.2711C50.5498 18.6789 49.4539 21.9336 49.1217 22.7396C51.1735 24.9806 52.4151 27.8352 52.4151 31.3289C52.4151 43.6197 44.929 46.3259 37.8033 47.1182C38.9511 48.1112 39.9738 50.0586 39.9738 53.0441C39.9738 57.3258 39.9367 60.7718 39.9367 61.8258C39.9367 62.6774 40.5128 63.6752 42.1348 63.361C54.8426 59.1249 64 47.1341 64 33.0003C64 15.3267 49.6728 1 32.0003 1Z"
+        fill={fillAreaColor}
+      />
+    </ComplicateIconBase>
+  );
+});
+GitHub.displayName = "GitHubIcon";
