+++ packages/toolkit/src/components/index.ts
@@ -1,3 +1,5 @@
+export * from "./top-bar";
+
 export * from "./AutoresizeInput";
 export * from "./Breadcrumb";
 export * from "./BreadcrumbWithLink";
@@ -23,5 +25,4 @@ export * from "./PageHead";
 export * from "./PageTitle";
 export * from "./cells";
 export * from "./TableError";
-export * from "./Topbar";
 export * from "./WarnUnsavedChangesModal";
