+++ packages/toolkit/src/components/top-bar/index.ts
@@ -1,0 +1,1 @@
+export * from "./Topbar";
